apps -> registration -> big int for cell num

+++ apps/registration/models.py
@@ -27,4 +27,4 @@ class Person(Customer):
     national_id = models.CharField(max_length=10, verbose_name="کد ملی")
     education = models.CharField(max_length=100, verbose_name="تحصیلات", null=True)
     org = models.CharField(max_length=100, verbose_name="موسسه/آموزشگاه/سازمان")
-    cellphone_number = models.IntegerField(verbose_name="شماره تلفن همراه")
+    cellphone_number = models.BigIntegerField(verbose_name="شماره تلفن همراه")
